Revert "Passage sur 4 octets pour la PI3"
This reverts commit 01c527716f12b79475b8f1a863d1416fd2747396.

+++ src/define.h
@@ -45,7 +45,7 @@
 #endif
 
 // Stockage de la valeurs des pulses compté
-extern volatile signed long nbEncoches;
+extern volatile signed int nbEncoches;
 
 // Pour la configuration de l'invertion des valeurs
 boolean invert;
